pyproject.toml
[project]
name = "dl-sound-classification"
version = "0.1.0"
description = "DL Sound Classification"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "hydra-core>=1.3.2",
    "isort>=6.0.1",
    "lightning>=2.5.2",
    "mlflow>=3.1.0",
    "numpy>=2.3.1",
    "psutil>=7.0.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "pyngrok",
    "timm",
    "python-dotenv",
    "optuna>=4.4.0",
    "optuna-dashboard>=0.19.0",
    "plotly>=6.2.0",
]
